rigl/requirements.txt
absl-py>=0.6.0
dopamine-rl
gin
numpy>=1.15.4
six>=1.12.0
tensorflow>=1.12.0,<2.0  # change to 'tensorflow-gpu' for gpu support
tensorflow-datasets==2.1
tensorflow-model-optimization
tf-agents
